refactor: inline preventLabelMultiClickSelect into Label component
It's only used by Label.svelte now, so the standalone module was
unnecessary indirection.
